lexer, parser: add more precision to floats

+++ src/lexing/lexer.rs
@@ -106,7 +106,7 @@ impl Lexer {
 
         match raw_number.parse::<i32>() {
             Ok(int) => self.push_token(TokenKind::Int(int)),
-            Err(_) => self.push_token(TokenKind::Float(raw_number.parse::<f32>().unwrap())),
+            Err(_) => self.push_token(TokenKind::Float(raw_number.parse::<f64>().unwrap())),
         }
     }
